requirements.txt
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
coverage==5.5
decorator==5.0.9
Flask==2.0.1
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
python-dotenv==0.19.0
ratelim==0.1.6
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
